add: health check

+++ deploy.sh
@@ -5,10 +5,16 @@ then
     docker cp $2 $1:/server/model.sav
     echo "restart $1 host"
     docker restart $1
-    echo "update nginx server"
-    docker cp nginx/$1.conf nginx:/etc/nginx/conf.d/server.conf
-    docker kill -s HUP nginx
-    echo "already switched to $1 host"
+    echo "health check"
+    if [ "$(docker exec -it $1 python3 healthCheck.py)" != "Error" ]
+    then
+        echo "update nginx server"
+        docker cp nginx/$1.conf nginx:/etc/nginx/conf.d/server.conf
+        docker kill -s HUP nginx
+        echo "already switched to $1 host"
+    else
+        echo "Health Check Failed"
+    fi
 else
     echo "Invalid"
 fi
